% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{ChooseFeatures}
\alias{ChooseFeatures}
\title{Choose features}
\usage{
ChooseFeatures(
  object,
  method,
  perturbation_meta = NULL,
  control_class = NULL,
  max_per_perturbation = 500,
  mean_expression = 0.01
)
}
\arguments{
\item{object}{A Seurat object}

\item{method}{"de" to choose features with differential expression; "expression" to choose based on expression levels}

\item{perturbation_meta}{If method = "de": the name of the metadata column containing the perturbation identities}

\item{control_class}{If method = "de": the name of the perturbation identity corresponding to the control condition}

\item{max_per_perturbation}{If method = "de": the maximum number of DEGs to use per perturbation (default 500)}

\item{mean_expression}{If method = "expression": the minimum average expression (default 0.01)}
}
\value{
A set of features to use for fingerprint estimation
}
\description{
Choose features for fingerprint estimation.
}
\examples{
\dontrun{
features <- ChooseFeatures(
object,
method = 'de',
perturbation_meta = 'perturbation_id',
control_class = 'NT'
)

features <- ChooseFeatures(
object,
method = 'expression'
)
}

}
