% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{ExplainMatch}
\alias{ExplainMatch}
\title{Explain match}
\usage{
ExplainMatch(
  object,
  query,
  match,
  condition_meta,
  control_class,
  dictionary,
  group_level,
  ref = NULL,
  perturbation_meta = NULL,
  control_class_ref = NULL,
  plot = TRUE,
  num_genes = 25,
  include_de = 0,
  query_key = NULL,
  ref_key = NULL,
  downsample_control = TRUE
)
}
\arguments{
\item{object}{A Seurat object}

\item{query}{Either a cell name or a group name}

\item{match}{The perturbation match to be explained; should be the name of one fingerprint in the dictionary}

\item{condition_meta}{The name of the metadata column containing the condition identities}

\item{control_class}{The name of the condition identity corresponding to the control condition}

\item{dictionary}{The dictionary of fingerprints that were used to map the data}

\item{group_level}{Whether fingerprinting was run at the group level (TRUE) or not (FALSE)}

\item{ref}{The Seurat object for the reference data from which the fingerprints were learned (only needed if plot is TRUE); if NULL, will use pseudobulked data stored in dictionary}

\item{perturbation_meta}{The name of the metadata column containing the perturbation identities in the reference data (only needed if plot is TRUE)}

\item{control_class_ref}{The name of the perturbation identity corresponding to the control condition in the reference data (only needed if plot is TRUE)}

\item{plot}{If TRUE (default for group-level), will return a heatmap comparison; otherwise, or for cell-level assignments, returns the top genes}

\item{num_genes}{Number of top genes to show (default 10)}

\item{include_de}{Number of DEGs from the reference and query each to show (default 0), only needed if plot is TRUE}

\item{query_key}{If different gene key types, the key type of the query object; NULL (default) if no conversion needed (only needed if plot is TRUE)}

\item{ref_key}{If different gene key types, the key type of the reference object; NULL (default) if no conversion needed (only needed if plot is TRUE)}

\item{downsample_control}{Whether the number of control cells should be downsampled (default TRUE), only needed if plot is TRUE}
}
\value{
If plot = TRUE, two heatmaps corresponding to the query and reference data explaining the match that was made; otherwise, a dataframe of the top genes responsible for a given match
}
\description{
Explain an assignment by identifying the top genes driving a match.
}
