% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{FingerprintCells}
\alias{FingerprintCells}
\title{Fingerprint cells}
\usage{
FingerprintCells(
  object,
  condition_meta,
  control_class,
  dictionary,
  group_level,
  top = 100,
  mc.cores = 1,
  seed.use = 1448145,
  suffix = NULL
)
}
\arguments{
\item{object}{A Seurat object}

\item{condition_meta}{The name of the metadata column containing the condition identities}

\item{control_class}{The name of the condition identity corresponding to the control condition}

\item{dictionary}{The dictionary of fingeprrints to map the data to}

\item{group_level}{Whether classification should be done for each cell, or for each condition in aggregate}

\item{top}{To speed up classification, maximum number of top fingerprints to include in regression (default 100); set NULL to use all}

\item{mc.cores}{The number of cores to use for parallelization (default 1)}

\item{seed.use}{Set a random seed (default 1448145); setting NULL will not set a seed}

\item{suffix}{Suffix for metadata columns; defaults to NULL under cell-level classification or the specified condition name if group-level}
}
\value{
A Seurat object with the following new metadata columns, appended by the specified suffix if any:
\describe{
\item{top_credible_set}{The top credible set assigned to each cell (or the condition to which the cell belongs).
Distinct members are separated by commas.}
\item{top_lbf}{The log Bayes factor associated with the top credible set.}
\item{all_credible_sets}{All credible sets assigned to each cell (or the condition to which the cell belongs).
Distinct sets are separated by semicolons, and distinct members within a credible set are separated by commas.}
\item{all_lbfs}{The log Bayes factors associated with each credible set, separated by semicolons.}
\item{all_signs}{The signs of association for each member (separated by commas) of each credible set (separated by semicolons).}
}
}
\description{
Map cells to fingerprints.
}
\examples{
\dontrun{
object <- FingerprintCells(
object,
condition_meta = 'condition_id',
control_class = 'untreated',
dictionary = dictionary,
group_level = FALSE
)

object <- FingerprintCells(
object,
condition_meta = 'condition_id',
control_class = 'untreated',
dictionary = dictionary,
group_level = TRUE
)
}

}
