% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{LearnDictionary}
\alias{LearnDictionary}
\title{Learn dictionary}
\usage{
LearnDictionary(
  object,
  perturbation_meta,
  control_class,
  features,
  min_cells = 10,
  override_filter = FALSE,
  mc.cores = 1,
  seed.use = 1448145
)
}
\arguments{
\item{object}{A Seurat object}

\item{perturbation_meta}{The name of the metadata column containing the perturbation identities}

\item{control_class}{The name of the perturbation identity corresponding to the control condition}

\item{features}{The features to be used for fingerprint estimation}

\item{min_cells}{The minimum number of cells that must be present to estimate a fingerprint (default 10)}

\item{override_filter}{If TRUE, low-signal fingerprints will not be filtered out (default FALSE)}

\item{mc.cores}{The number of cores to use for parallelization (default 1)}

\item{seed.use}{Set a random seed (default 1448145); setting NULL will not set a seed}
}
\value{
A list of fingerprints and their distributions for downstream fingerprinting analysis:
\describe{
\item{Lambdas}{The matrix of fingerprints corresponding to each perturbation}
\item{dists}{A list of means and standard deviations for each entry of each fingerprint}
}
}
\description{
Learn dictionary of fingerprints corresponding to each perturbation.
}
\examples{
\dontrun{
dictionary <- LearnDictionary(
object,
perturbation_meta = 'perturbation_id',
control_class = 'NT',
features = features
)
}

}
