% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{LongIslandCityPlot}
\alias{LongIslandCityPlot}
\title{Long Island City plot}
\usage{
LongIslandCityPlot(
  object,
  query,
  condition_meta,
  control_class,
  dictionary,
  group_level,
  suffix = NULL,
  approx = FALSE,
  order = "clustering",
  cs = 10
)
}
\arguments{
\item{object}{A Seurat object}

\item{query}{Either a cell name or a group name}

\item{condition_meta}{The name of the metadata column containing the condition identities}

\item{control_class}{The name of the condition identity corresponding to the control condition}

\item{dictionary}{The dictionary of fingerprints that the data was mapped to}

\item{group_level}{Whether fingerprinting was run at the group level (TRUE) or not (FALSE)}

\item{suffix}{The suffix of the output metadata columns; defaults to NULL if cell-level or the specified condition name if group-level}

\item{approx}{If TRUE, plot marginal correlations with fingerprints; if FALSE (default), plot conditional posterior mean coefficients}

\item{order}{Whether to order fingerprints by clustering (default) or alphabetically; ignored if order is stored in fingerprints}

\item{cs}{The maximum number of top credible sets to visualize (default 1)}
}
\value{
A Long Island City plot visualizing assignment of the query
}
\description{
Visualize assignment with a Long Island City plot.
}
\examples{
\dontrun{
LongIslandCityPlot(
object,
query = 'R1-L1_AAAGGATCACCAGCTG',
condition_meta = 'condition_id',
control_class = 'untreated',
dictionary = dictionary,
group_level = FALSE,
cs = 1
)

LongIslandCityPlot(
object,
query = 'drug1',
condition_meta = 'condition_id',
control_class = 'untreated',
dictionary = dictionary,
group_level = TRUE,
cs = 10
)
}

}
