% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{ProjectFingerprints}
\alias{ProjectFingerprints}
\title{Project fingerprints}
\usage{
ProjectFingerprints(
  object,
  condition_meta,
  control_class,
  dictionary,
  suffix = NULL
)
}
\arguments{
\item{object}{A Seurat object}

\item{condition_meta}{The name of the metadata column containing the condition identities}

\item{control_class}{The name of the condition identity corresponding to the control condition}

\item{dictionary}{The dictionary of fingerprints that were used to map the data}

\item{suffix}{The suffix of the output metadata columns; defaults to NULL}
}
\value{
A Seurat object with a 'proj' assay containing projected values and 'projumap' reduction
}
\description{
Project assigned cells to the fingerprints and produce a UMAP visualization.
}
\examples{
\dontrun{
object <- ProjectFingerprints(
object,
condition_meta = 'condition_id',
control_class = 'untreated',
dictionary = dictionary
)
}

}
