% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{SummarizeResults}
\alias{SummarizeResults}
\title{Summarize results}
\usage{
SummarizeResults(
  object,
  group_level,
  condition_meta,
  control_class,
  suffix = NULL,
  tidy = FALSE
)
}
\arguments{
\item{object}{A Seurat object}

\item{group_level}{Whether fingerprinting was run at the group level (TRUE) or not (FALSE)}

\item{condition_meta}{The name of the metadata column containing the condition identities}

\item{control_class}{The name of the perturbation identity corresponding to the control condition}

\item{suffix}{The suffix of the output metadata columns; defaults to NULL if cell-level or the specified condition name if group-level}

\item{tidy}{Whether results should be in tidy format (default FALSE)}
}
\value{
A dataframe summarizing the mapping results
}
\description{
Summarize results after mapping cells to fingerprints.
}
\examples{
\dontrun{
summary <- SummarizeResults(
object,
group_level = TRUE,
condition_meta = 'perturbation_id',
control_class = 'NT'
)

summary <- SummarizeResults(
object,
group_level = FALSE,
condition_meta = 'perturbation_id',
control_class = 'NT'
)

summary <- SummarizeResults(
object,
group_level = FALSE,
condition_meta = 'perturbation_id',
control_class = 'NT',
tidy = TRUE
)
}

}
